pyproject.toml
[project]
name = "prompting-workshop"
version = "0.0.1"
description = "This is a template repository for Python projects that use uv for their dependency management."
authors = [{ name = "Anis Zakari", email = "anis.zakari@outlook.fr" }]
readme = "README.md"
keywords = ['python']
requires-python = ">=3.10,<4.0"
classifiers = [
    "Intended Audience :: Developers",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "gradio>=5.29.0",
    "jupyter>=1.1.1",
    "langchain>=0.3.24",
    "langchain-community>=0.3.23",
    "langchain-openai>=0.3.14",
    "langserve[all]>=0.3.1",
    "langsmith>=0.3.34",
    "loguru>=0.7.3",
    "pyngrok>=7.2.8",
    "python-dotenv>=1.1.0",
    "rank-bm25>=0.2.2",
    "rich>=14.0.0",
    "uvicorn>=0.34.2",
]

[project.urls]
Homepage = "https://AnisZakari.github.io/prompting-workshop/"
Repository = "https://github.com/AnisZakari/prompting-workshop"
Documentation = "https://AnisZakari.github.io/prompting-workshop/"

[dependency-groups]
dev = [
    "pytest>=7.2.0",
    "pre-commit>=2.20.0",
    "tox-uv>=1.11.3",
    
    "mypy>=0.991",
    
    "ruff>=0.11.5",
    
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.setuptools]
py-modules = ["prompting_workshop"]

[tool.mypy]
files = ["prompting_workshop"]
disallow_untyped_defs = true
disallow_any_unimported = true
no_implicit_optional = true
check_untyped_defs = true
warn_return_any = true
warn_unused_ignores = true
show_error_codes = true

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
target-version = "py39"
line-length = 120
fix = true

[tool.ruff.lint]
select = [
    # flake8-2020
    "YTT",
    # flake8-bandit
    "S",
    # flake8-bugbear
    "B",
    # flake8-builtins
    "A",
    # flake8-comprehensions
    "C4",
    # flake8-debugger
    "T10",
    # flake8-simplify
    "SIM",
    # isort
    "I",
    # mccabe
    "C90",
    # pycodestyle
    "E", "W",
    # pyflakes
    "F",
    # pygrep-hooks
    "PGH",
    # pyupgrade
    "UP",
    # ruff
    "RUF",
    # tryceratops
    "TRY",
]
ignore = [
    # LineTooLong
    "E501",
    # DoNotAssignLambda
    "E731",
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101"]

[tool.ruff.format]
preview = true


